+++ client/src/App.js
@@ -1,3 +1,4 @@
+import { useEffect, useState } from 'react';
 import './styles/App.css'
 import Login from './pages/Login';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
@@ -5,19 +6,19 @@ import DirecteurRoutes from './routes/DirecteurRoutes';
 import NoteFound from './pages/NoteFound';
 
 function App() {
-  const role = JSON.parse(localStorage.getItem('token')).fonction;
+  const [token, setToken] = useState(null)
+
+  useEffect(() => {
+    setToken(JSON.parse(localStorage.getItem('token')))
+  }, [])
+
   return (
     <div className="App">
       <BrowserRouter>
         <Routes>
-          <Route path="/" element={<Login />} />
-          {role === 'formateur' ? 
-            <>
-              <Route path="/directeur/*" element={<DirecteurRoutes />} />
-            </>
-            :
-            <Route path='*' element={<NoteFound/>} />
-          }
+          <Route path="/" element={<Login />} /> 
+          <Route path="/directeur/*" element={<DirecteurRoutes />} />
+          <Route path='*' element={<NoteFound/>} />
         </Routes>
       </BrowserRouter>
     </div>
