--- package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "papaparse": "^5.4.1"
-  }
-}
